requirements.txt
fastapi
uvicorn
paddleocr3.3.2
paddlepaddle
pymongo3.12
python-multipart
python-dotenv